refactor: 사다리 높이가 스스로 자연수임을 검증 구현
