+++ src/test/java/laddergame/domain/LadderTest.java
@@ -15,14 +15,14 @@ public class LadderTest {
     @DisplayName("생성에 성공한다.")
     public void createLadderTest() {
         //given
-        final String height = "4";
+        final Height height = new Height("4");
         final int playerCount = 4;
 
         //when
         Ladder ladder = new Ladder(playerCount, height);
 
         //then
-        assertEquals(ladder.getLines().size(), Integer.parseInt(height));
+        assertEquals(ladder.getLines().size(), height.getHeight());
         assertEquals(ladder.getLines().get(0).getPoints().size(), playerCount - 1);
     }
 
@@ -30,7 +30,7 @@ public class LadderTest {
     @DisplayName("사다리 다리 건설 생성에 성공하는지 테스트한다.")
     public void buildLadderBridge() {
         //given
-        final String height = "1";
+        final Height height = new Height("1");
         final int playerCount = 4;
         List<LineBuildResult> buildResult = new ArrayList<>();
         CanBuildStrategy canBuildStrategy = new CanBuildStrategy() {
